🌱 split compute: extract data into separate module
